feat: Dedicated signup service
