+++ config/web/auth.php
@@ -2,8 +2,13 @@
 
 declare(strict_types=1);
 
+use App\Auth\Identity;
+use App\Auth\IdentityRepository;
+use Cycle\ORM\ORMInterface;
+use Psr\Container\ContainerInterface;
 use Psr\Log\LoggerInterface;
 use Yiisoft\Access\AccessCheckerInterface;
+use Yiisoft\Auth\IdentityRepositoryInterface;
 use Yiisoft\Definitions\Reference;
 use Yiisoft\Cookies\CookieEncryptor;
 use Yiisoft\Cookies\CookieMiddleware;
@@ -15,6 +20,12 @@ use Yiisoft\User\Login\Cookie\CookieLogin;
 /** @var array $params */
 
 return [
+    IdentityRepositoryInterface::class => static function (ContainerInterface $container): IdentityRepository {
+        return $container
+            ->get(ORMInterface::class)
+            ->getRepository(Identity::class);
+    },
+
     CookieMiddleware::class => static fn (CookieLogin $cookieLogin, LoggerInterface $logger) => new CookieMiddleware(
         $logger,
         new CookieEncryptor($params['yiisoft/cookies']['secretKey']),
